add suffle_test.cpp test_utils.h

+++ omnioperator/omniop-spark-extension/cpp/src/shuffle/splitter.h
@@ -31,7 +31,7 @@ using namespace omniruntime::mem;
 struct SplitRowInfo {
     uint32_t copyedRow = 0;
     uint32_t onceCopyRow = 0;
-    uint32_t remainCopyRow = 0;
+    uint64_t remainCopyRow = 0;
     vector<uint32_t> cacheBatchIndex; // 记录各定长列的溢写Batch下标
     vector<uint32_t> cacheBatchCopyedLen; // 记录各定长列的溢写Batch内部偏移
 };
@@ -74,7 +74,7 @@ class Splitter {
 
     std::vector<int32_t> partition_id_; // 记录当前vb每一行的pid
     std::vector<int32_t> partition_id_cnt_cur_; // 统计不同partition记录的行数(当前处理中的vb)
-    std::vector<int32_t> partition_id_cnt_cache_; // 统计不同partition记录的行数，cache住的
+    std::vector<uint64_t> partition_id_cnt_cache_; // 统计不同partition记录的行数，cache住的
     // column number
     uint32_t num_row_splited_; // cached row number
     uint64_t cached_vectorbatch_size_; // cache total vectorbatch size in bytes
